+++ src/MessageHandler/FeedNormalizationHandler.php
@@ -12,15 +12,11 @@ use Symfony\Component\Messenger\MessageBusInterface;
 #[AsMessageHandler]
 final class FeedNormalizationHandler
 {
-    /**
-     * Max length here is taken from the max database varchar length.
-     */
-    private const EXCERPT_MAX_LENGTH = 255;
-
     public function __construct(
         private readonly ContentNormalizer $contentNormalizer,
         private readonly MessageBusInterface $messageBus,
         private readonly TagsNormalizerInterface $tagsNormalizer,
+        private readonly int $excerptMaxLength,
     ) {
     }
 
@@ -38,7 +34,7 @@ final class FeedNormalizationHandler
         $item->description = $this->contentNormalizer->normalize($item->description ?? '');
         if (!is_null($item->excerpt)) {
             $item->excerpt = $this->contentNormalizer->normalize($item->excerpt);
-            $item->excerpt = $this->contentNormalizer->trimLength($item->excerpt, self::EXCERPT_MAX_LENGTH);
+            $item->excerpt = $this->contentNormalizer->trimLength($item->excerpt, $this->excerptMaxLength);
         }
 
         $this->messageBus->dispatch(new EventMessage($item));