+++ CHANGELOG.md
@@ -33,6 +33,7 @@ See [keep a changelog] for information about writing changes to this log.
 - Added indexing service and helper commands to populate and create indexes
 - Added Easy admin and event fixtures
 - Make data imported from feeds read-only in easy admin
+- Added event subscriber to index content created in the UI
 
 [keep a changelog]: https://keepachangelog.com/en/1.1.0/
 [unreleased]: https://github.com/itk-dev/event-database-imports/compare/main...develop